requirements.txt
asgiref==3.2.3
confusable-homoglyphs==3.2.0
Django==3.0.3
django-crispy-forms==1.8.1
django-favicon==0.1.3
django-markdown-deux==1.0.5
django-pagedown==2.0.3
django-registration==3.1
djangorestframework==3.11.0
gunicorn==20.0.4
markdown2==2.3.8
Pillow==7.0.0
pytz==2019.3
sqlparse==0.3.0
